--- a/Directory.Build.props
+++ b/Directory.Build.props
@@ -42,6 +42,7 @@
   <ItemGroup>
     <Using Include="Microsoft.Toolkit.HighPerformance" />
     <Using Include="System.Drawing" />
+    <Using Include="System.Numerics" />
     <Using Include="System.Globalization" />
     <Using Include="System.Text" />
   </ItemGroup>