--- a/benchmark.ps1
+++ b/benchmark.ps1
@@ -5,7 +5,7 @@
 
 param(
     [Parameter(Mandatory = $false)][string] $Configuration = "Release",
-    [Parameter(Mandatory = $false)][string] $Framework = "net6.0",
+    [Parameter(Mandatory = $false)][string] $Framework = "net7.0",
     [Parameter(Mandatory = $false)][string] $Filter = "*",
     [Parameter(Mandatory = $false)][string] $Job = "short"
 )