--- a/src/AdventOfCode/AdventOfCode.csproj
+++ b/src/AdventOfCode/AdventOfCode.csproj
@@ -12,6 +12,5 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Toolkit.HighPerformance" />
-    <PackageReference Include="System.Runtime.Experimental" />
   </ItemGroup>
 </Project>