--- a/src/AdventOfCode/Puzzles/Y2017/Day07.cs
+++ b/src/AdventOfCode/Puzzles/Y2017/Day07.cs
@@ -124,7 +124,7 @@ public sealed class Day07 : Puzzle
         }
 
         var unbalancedSubtree = root.Children
-            .GroupBy((p) => FindDesiredWeightOfUnbalancedDisc(p))
+            .GroupBy(FindDesiredWeightOfUnbalancedDisc)
             .Where((p) => p.Key is not 0)
             .SingleOrDefault();