--- a/src/AdventOfCode/Puzzles/Y2017/Day12.cs
+++ b/src/AdventOfCode/Puzzles/Y2017/Day12.cs
@@ -36,7 +36,7 @@ public sealed class Day12 : Puzzle
         {
             var members = GetMembersOfGroup(target);
 
-            string groupId = string.Join(',', members.Select((p) => p.Id).OrderBy((p) => p));
+            string groupId = string.Join(',', members.Select((p) => p.Id).Order());
             groups.Add(groupId);
         }