… .vscode/launch.json
… global.json
… src/AdventOfCode.Site/Properties/launchSettings.json
… src/AdventOfCode.Site/aws-lambda-tools-defaults.json
… src/AdventOfCode.Site/package.json
… src/AdventOfCode/Puzzles/Y2020/Day20.cs
… src/AdventOfCode/Puzzles/Y2020/Day22.cs
… src/AdventOfCode/Puzzles/Y2021/Day09.cs
… src/AdventOfCode/Puzzles/Y2021/Day17.cs
… src/AdventOfCode/StringExtensions.cs
… startvs.cmd
… startvscode.cmd
… tests/AdventOfCode.Benchmarks/AdventOfCode.Benchmarks.csproj
… tests/AdventOfCode.Tests/AdventOfCode.Tests.csproj
… tests/AdventOfCode.Tests/Api/ApiTests.cs
… tests/AdventOfCode.Tests/Api/ResourceTests.cs
… tests/AdventOfCode.Tests/EndToEnd/ResourceTests.cs