Update to ASP.NET Core 7 release candidate 1.