@martincostello 2022-02-19
```md
| benchmark                  | mean (microbenchmarks-2021.base) | mean (microbenchmarks-2021.pr) | ratio | allocated (microbenchmarks-2021.base) | allocated (microbenchmarks-2021.pr) | ratio |
| -------------------------- | -------------------------------- | ------------------------------ | ----- | ------------------------------------- | ----------------------------------- | ----- |
| Solve2021(input: Y2021-01) |                         190.3 µs |                       191.0 µs |  1.00 |         234.64 KB |                           234.64 KB |  1.00 |
| Solve2021(input: Y2021-02) |                         221.0 µs |                       222.4 µs |  1.01 |         246.94 KB |                           246.94 KB |  1.00 |
| Solve2021(input: Y2021-03) |                         341.8 µs |                       347.2 µs |  1.02 |         161.08 KB |                           161.08 KB |  1.00 |
| Solve2021(input: Y2021-04) |                         908.6 µs |                       925.3 µs |  1.02 |         353.86 KB |                           353.86 KB |  1.00 |
| Solve2021(input: Y2021-05) |                          29.9 ms |                        30.4 ms |  1.02 |          16.76 MB |                            16.76 MB |  1.00 |
| Solve2021(input: Y2021-06) |                         116.9 µs |                       117.2 µs |  1.00 |          20.89 KB |                            20.89 KB |  1.00 |
| Solve2021(input: Y2021-07) |                          33.9 ms |                        33.1 ms |  0.98 |         170.22 KB |                           170.22 KB |  1.00 |
| Solve2021(input: Y2021-08) |                           3.3 ms |                         3.2 ms |  0.97 |           2.54 MB |                             2.54 MB |  1.00 |
| Solve2021(input: Y2021-09) |                           6.3 ms |                         6.3 ms |  1.00 |           3.03 MB |                             3.03 MB |  1.00 |
| Solve2021(input: Y2021-10) |                         151.7 µs |                       153.6 µs |  1.01 |          53.19 KB |                            53.19 KB |  1.00 |
| Solve2021(input: Y2021-11) |                           7.9 ms |                         8.1 ms |  1.02 |           2.58 MB |                             2.58 MB |  1.00 |
| Solve2021(input: Y2021-12) |                         112.4 ms |                       107.2 ms |  0.95 |          19.28 MB |                            19.28 MB |  1.00 |
| Solve2021(input: Y2021-13) |                         712.4 µs |                       712.2 µs |  1.00 |         492.55 KB |                           492.55 KB |  1.00 |
| Solve2021(input: Y2021-14) |                         700.5 µs |                       714.6 µs |  1.02 |         592.51 KB |                           592.51 KB |  1.00 |
| Solve2021(input: Y2021-15) |                         273.2 ms |                       251.1 ms |  0.92 |          60.71 MB |                            60.71 MB |  1.00 |
| Solve2021(input: Y2021-16) |                          91.4 µs |                        85.1 µs |  0.93 |         198.82 KB |                           198.82 KB |  1.00 |
| Solve2021(input: Y2021-17) |                          11.2 ms |                        11.4 ms |  1.01 |          21.65 MB |                            21.65 MB |  1.00 |
| Solve2021(input: Y2021-18) |                         217.0 ms |                       240.1 ms |  1.11 |         239.77 MB |                           239.77 MB |  1.00 |
| Solve2021(input: Y2021-19) |                            2.0 s |                          2.0 s |  1.01 |           1.42 GB |                             1.42 GB |  1.00 |
| Solve2021(input: Y2021-20) |                         569.7 ms |                       586.7 ms |  1.03 |         102.55 MB |                           102.55 MB |  1.00 |
| Solve2021(input: Y2021-21) |                         251.8 ms |                       248.7 ms |  0.99 |         381.19 MB |                           381.19 MB |  1.00 |
| Solve2021(input: Y2021-22) |                         317.5 ms |                       345.3 ms |  1.09 |         140.81 MB |                           140.81 MB |  1.00 |
| Solve2021(input: Y2021-23) |                         906.4 ms |                       947.2 ms |  1.04 |         642.66 MB |                           642.66 MB |  1.00 |
| Solve2021(input: Y2021-24) |                          33.6 µs |                        32.6 µs |  0.97 |          42.81 KB |                            42.81 KB |  1.00 |
| Solve2021(input: Y2021-25) |                          78.1 ms |                        79.3 ms |  1.01 |             37 MB |                               37 MB |  1.00 |
```